refactor: move purchaseField to its own file
